MB-52067: Ensure bgfetch does not short circuit multi-phase ops
Increment is implemented with a SteppableCommandContext.
If any phase receives status `would_block`, the subsequent
notify_io_complete call _must_ provide `success` if the steppable
context needs to run again.

Any other status will result in a corresponding response being written
and any remaining stages of the command context being skipped.

If operations overlap, a bgfetch may complete but not find a temporary
item in the HashTable (also possible if the paging visitor removes the
temp item). Previously, if the item was not found on disk this could
lead to the cookie being notified with no_such_key, potentially ending
a command with an unexpected no_such_key response.

To avoid this specific case, this patch ensures bgfetches which:

* don't find the item on disk
* don't find a temp item in the HT

notifies the cookie with success.

This is not entirely new behaviour - if a temp item _is_ found, the
cookie is already notified with success, relying on the operation
to re-run and find the temp non existent item.

This does introduce inefficiency - multiple concurrent GETs for the
same key might now wastefully trigger repeated bgfetches.
To resolve this, a future patch _may_ prevent GETs cleaning up temp
items (as the ItemPager will do so eventually).

Change-Id: I825739c3ccdb373850a40f442b3e829e7d65080d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/175871
Well-Formed: Restriction Checker
Tested-by: James H <james.harrison@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>