[BP] MB-52276: Fix CheckpointCursor::getRemainingItemsCount
In previous patches Checkpoint::numItems has been changed from
accounting only non-meta items to accounting all items in the
checkpoint (empty item excluded).
After that change CheckpointCursor::getRemainingItemsCount() provides a
wrong/hybrid count of meta/non-meta items, so we fix that here.
With this patch "remaining items for cursor" accounts for all items
remaining, same as Checkpoint::numItems.

Note: A collateral effect of that is that cursors in checkpoints like

    [e cs vbs)
     ^

have remaining_items_count > 0 for DCP streams.
That is a problem for users that rely on that DCP stat for inferring
whether replication has settled (see MB-52553 for details).

Thus, this patch also ensures that DCP streams move their cursor to
the end of checkpoint in that state.
ActiveStream is resilient to that. This patch adds specific test
coverage for that.

Note: The fix here allows to easily make
CheckpointCursor::getRemainingItemsCount O(1). Dedicated follow-up for
that.

Change-Id: I1634cc83290a08ec8e3c934c343653ee24960d98
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194212
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>