MB-52504: Make Collector responsible for "kv_" metric prefix
Rather than being a property of the endpoint (e.g., _prometheusMetrics,
_metering), manage the prefix in the PrometheusStatCollector.

This is to facilitate having mixed prefixed and non-prefixed metrics in
one endpoint, to allow the local Prometheus instance to gather metering
metrics without specifically scraping _metering (see MB-52123 for why
this is needed). This will be added in a later patch.

A matching change will also be made by other components.

Change-Id: If8e9960e746f15540c5c25f8798d0e1531054d4d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176159
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>