MB-51606: Change ADM::getTrackedWrites() to getTrackedKeys()
We only need the keys in the place in which we use it, the DM
logic only checks for committed states. This allows us to
re-use the function name (getTrackedWrites()) to return SyncWrite
objects which we will use to transition to ADM/PDM with a generic DM
in a future change.

Change-Id: Ic69990f8cdd760c81df675173d42025b4b3e3286
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176180
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>