MB-51606: Add DeadDurabilityMonitor
Add a new type of DurabilityMonitor that we create and use for
vBuckets in the dead state. Currently when we transition to a
dead state vBucket we keep the current DurabilityMonitor (ADM or
PDM). This leads to confusion, and potentially error prone code
as it may not be valid to do certain things with ADMs or PDMs for
a vBucket in the dead state. This allows us to deal with
notifications such as persistence in the DurabilityMonitor of a
vBucket in the dead state in a graceful manner in a future change.

Change-Id: Ic2b8505fa56cd53cf4d1b2bc0f2a4d7af0d56f8a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176192
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>