MB-52264: Remove "max_size" configuration listener
The "max_size" configuration parameter determines the Bucket quota.
Currently when it is changed we update the HWM, LWM, and other
related values. The BucketQuotaChangeTask is going to own this
work in a future change, so in preparation for that remove the
configuration change listener and have it call the current function
directly.

Change-Id: I71b2950b8cfa3b837c89583ad1529361ffe7960f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176365
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>