Fix memory tracking for memcached on OSX
All test programs were unaffected by what is suspected is a change
in linker behaviour, hence the issue goes undetected. It seems that
the way that the platform_cb_malloc_arena (weak symbols) are linked
into memcached via memcached_daemon no longer results in those symbols
being present in the final memcached binary.

This is addressed by changing the linker definitions and the addition
of a run-time check that we are reaching the correct cb_malloc backend.

Change-Id: I5315917311cb88291d212ff2e8f57d5dc6bbf74a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176308
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>