MB-35297: Make RangeScan config duration and buffering dynamic
Add "set" paths for changing the runtime value of:

* range_scan_read_buffer_send_size

This value is read from the config when a new scan is created, existing
scans do not pick-up the change.

* range_scan_max_lifetime

This value is push through by a change-listener to the range scan code.
New values take affect when the timeout task is scheduled.

These are deliberately not documented in cbepctl

Note: this patch includes an unrelated range_scan_create.md fix to
avoid CV on a standalone doc patch.

Change-Id: I578ba6dd35fe6859667468473a53159445d9e167
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176433
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>