MB-52568: Print actual HPS for ADM's "high_prepared_seqno" stat
The stat named "high_prepared_seqno" in the durability-monitor stat
group is not strictly the HPS, it's actually the seqno which has been
ack'd locally by the active node, out of the SyncWrites which are
tracked.

This is potentially misleading as it can differ from the actual "HPS"
value - for example Immediately after warmup (before any more
SyncWrites are performed), there will not be any SyncWrites tracked
yet, and hence this quantity returns zero.

Given we already print the value of the active's last_write_seqno via
the chain information, for the "high_prepared_seqno" stat simply print
the HPS value to avoid confusion.

Change-Id: I313fa5903be7ee5624cadfd31e9513f9584201d3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176445
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>