MB-52553: Revert "MB-52276: Fix CheckpointCursor::getRemainingItemsCount"
This reverts commit a3eb28cd7916390a61a027e2707c332673cc55f0.

The original patch makes remaining-items-count for a cursor to account
also non-meta items in a checkpoint.

If that count is taken for when a cursor is placed into a "empty" open
checkpoint, then dcp_items_remaining would never drop to 0 as a DCP
doens't move away from the dummy item in that state.
That is currently failing QE/Perf tests that rely on
"dcp_items_remaining = 0" for inferring the end of some steps in their
tests.

I believe that the fix for that problem is allowing that cursor-move.
A change is in-progress for implementing that change, with full
test-coverage that verifies that the checkpoint_start processing is
not lost with the new behaviour. Patch at
https://review.couchbase.org/c/kv_engine/+/176513.
But, that patch currently causes some failures in a couple of takeover
tests that require further investigation.

This is a quick-fix for unblocking QE/Perf on their work.

Change-Id: I39819df877320eb312ceb5b7792c230a3e0240d0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176541
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>