MB-52699: Check for an error when Magma-KVStore::scan loop ends
Both seq and id iterators lack a GetStatus check with the iterator
is no longer Valid. It could be that a critical error has occurred
causing the loop to end early.

No testing in this commit as there is no suitable framework for forcing
an error in magma.

Change-Id: I3c04c4695e825928994ffb923f6f1bbcb229d8d1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176643
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Restriction Checker