MB-35297: Add a byte limit to range-scan-continue
Add the byte limit to the range-scan-continue command, allowing
a client to set a limit on how much data can be sent before the
range-scan-continue ends. As each range-scan-continue will always
return at least one key/document, the value is not a hard boundary
and a range-scan-continue will go over the limit for only one
key/document.

On the wire the limit is specified as a 4-byte value giving a plentiful
4GB boundary.

Change-Id: I1c1da3e56daceaa13d337b093e4ed20854adc2e7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176732
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>