MB-52784: Remove Cookie::error_json member variable
We do not need to maintain the serialised error JSON as a member of
Cookie, as we copy it into the response packet anyway when sent.

Remove the member variable and just pass ownership of the string to
the callers of getErrorJson().

Reduces the size of Cookie from 568 to 536 Bytes.

Change-Id: Ia5194e14bde2a174dc319454bdd225fcb34ee1ef
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176811
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>