RangeScan should be the last privilege
The PrivilegeMask is a bitset containing one bit per privilege
and gets sized by the value of the last privilege. We've had
cases where new privileges was added, but that privilege mask
wasn't resized so one would try to reference invalid memory.

Add a comment that new privileges should be added _before_
the RangeScan privilege to try to avoid this to happen
again.

Change-Id: I7c548dabb80776b6a340892be1cbfd9bcb722885
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176809
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>