MB-52784: Outline Cookie::euid and Cookie::euidPrivilegeContext
Cookie member variables euid and euidPrivilegeContext are optional
fields which are only used if the 'impersonate' flags are set on a
request.

They are currently implemented using std::optional; which contains an
inline copy of the Type (along with a boolean flag indicating if the
opional has a value). This results in the follow memory footprint in
Cookie:

    /*  384      |    48 */    class std::optional<cb::rbac::UserIdent> {

				   /* total size (bytes):   48 */
			       } euid;
    ...
    /*  448      |    48 */    class std::optional<cb::rbac::PrivilegeContext> {

				   /* total size (bytes):   48 */
			       } euidPrivilegeContext;

Given that the vast majority of requests to not use impersonate,
outline these two members by using std::unique_ptr to implement
"optional-ness" - if/when impersonate is used we can allocate their
objects on demand. This does increase the cost of impersonate
fractionally (additional heap alloc plus pointer dereference), but
this is not expected to be performance-sensitive functionality.

Results in the following layout for euid & euidPrivilegeContext:

    /*  384      |     8 */        struct std::__uniq_ptr_data<_Tp, _Dp, true, true> [with _Tp = _Tp, _Dp = _Dp] : public std::__uniq_ptr_impl<_Tp, _Dp> {

				       /* total size (bytes):    8 */
				   } _M_t;

				   /* total size (bytes):    8 */
			       } euid;

    /*  408      |     8 */        struct std::__uniq_ptr_data<_Tp, _Dp, true, true> [with _Tp = _Tp, _Dp = _Dp] : public std::__uniq_ptr_impl<_Tp, _Dp> {

				       /* total size (bytes):    8 */
				   } _M_t;

				   /* total size (bytes):    8 */
			       } euidPrivilegeContext;

Overall reduces Cookie size from 520 to 440 Bytes.

Change-Id: Id555aa8d10dedd0df4126d852c2ca05edbaf9d48
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176814
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>