MB-52793: Ensure StoredValue::del updates datatype
If a StoredValue is deleted, but does not have a resident value (but
_may_ have one on disk, containing xattrs), previously `del`
erroneously skipped updating the datatype.

This situation has only been observed to occur on replicas, via
PassiveStream calling deleteWithMeta for an already deleted item. This
may occur when xattrs are removed from a deleted document.

See MB for more details.

Change-Id: I213cefb3907c4e290c2857c8526477f69a9ce764
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177197
Well-Formed: Restriction Checker
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>