MB-44253: Test: Should crash if GlobalTask::run throws
Add a regression test for MB-44253 - tasks which (incorrectly) throw
exceptions result in the process getting terminated (uncaught
exception).

Note that GlobalTasks *shouldn't* throw, but if they do we want to
treat that as a fatal error as there isn't a good way to handle
background tasks failing otherwise.

To achieve this with Folly we needed to patch it to not catch
exceptions in it's runTask() methods.

Change-Id: I5b92e937c55dd360db137de916705c7a66c0fc90
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176929
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>