MB-52728: StatsCommandContext: Handle bucket_get_stats() returning EWouldblock
Verify that the background tasks which perform budket-level STAT
requests correctly handle async notificaiton - prior to the fix for
this MB we could end up calling notifyIoComplete *twice* for a single
engine API call (there should only be one).

This manifested as intermittent failures in subsequent commands if the
same Cookie was reused - which was typically the case if running in
in-order mode.

Change-Id: Ib442600abfacffd821c8b3db94d7690f807c7b44
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177029
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>