MB-52728: Refactor: remove ServerCookieIface::execution_complete
This was added for RangeScan support, as that transmits all response
packets from the RangeScanDataHandler class inside ep-engine and hence
we didn't _need_ to execute() the Cookie a final time to wrap up.

However, having two methods which can wake up the front-end (the
original notify_io_complete() and the new execution_complete() method
complicates the API (which I'm trying to tighten up for EWouldBlock
handling).

We can actually use the existing notify_io_complete() call, by passing
the sent status (range_scan_more/range_scan_compelte) to it and then
checking for that status in range_scan_continue_executor, and simply
doing nothing in that case.

Change-Id: I7b92f9a068b1da222faa104fc1dcdc828307572d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177020
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>