MB-46738: Simplify FlowControl locking mechanism
The FlowControl bufferSize and pendingControl members are strictly
related, as any change to the first requires the Consumer sending a
control message to the Producer for informing the latter of the new
FlowControl setup.

Here I replace the manual locking/unlocking and pair-updates of those
members with a folly::Synchronized<T>, where T is a class that embeds
the pair-updates internally. That makes the FlowControl state update
cleaner.

Change-Id: I1e320f74894f92bb381fad82ba3b6deeb2229d08
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177017
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>