[BP] MB-52668 Count number of backfill pauses
Count of backfill pauses/resumes was earlier determined by the restarted
count log message in Magma's SeqIterator. For example:

    Ending seqno iterator (973409=>976673) with seqno 976673 restarted
    238 times

We've changed the log level for this message to debug as SeqIterator
will now very frequently be constructed/destructured after commit
kv-engine/d84225fc.

The count gives an idea of the memory pressue due to which backfills are
paused. To track this we introduce a counter in ActiveStream that is
incremented upon every backfill pause and is logged when the backfill is
completed.

Change-Id: Ic175fd8074024081f37f0c12dd56ffa24655da62
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177018
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/193437
Well-Formed: Restriction Checker