MB-52067: BGFetcher return success for success or no_such_key
If completeBGFetchForSingleItem returns no_such_key, consider it
unconditionally to be successful, this notifyIOComplete will always
see success allowing the front-end worker to revisit the engine and
determine how to proceed.

Note in terms of the MB triggering this fix, kv-engine generally never
needed this change because the VBucket owned bloomfilter decided if
the get would be successful, bypassing the need for a bg-fetch at all.
Now with magma as the KVStore, the VBucket owned bloomfilter isn't
enabled, hence we reach the bg-fetcher and a combination of bg-fetch
and hash-table visit are required to decide if a key exists or not.

Change-Id: I63aa7566d07466fdf37b3c61397490f43e8a57cd
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177038
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>