MB-52843: Skip ep_kv_size check in ep_testsuite 'disk>RAM golden path'
ep_kv_size consists of the memory footprint all Blobs + StoredValue
objects. As part of test 'disk>RAM golden path' it attempts to read an
initial value of ep_kv_size, and then compare that as the test
progresses and evicts / restores the single stored Document.

However this approach is flawed, as other non-visible meta items such
as SetVBucketState also allocate Blob objects. Depending on when
ep_kv_size is initially read, there may be /two/ Blob objects in
existance- one from the SetVBucketState and one from document "k1".

As such, when kv_size3 is later read, it can end up being a smaller
value than kv_size as the SetVBucketState Blob has been discarded
(when it was flushed to disk).

We have a number of unit tests which cover the specifics of memory
usage as items are added / removed from a VBucket, so address this
issue by skipping the checks for ep_kv_size.

Change-Id: Ib2cbd937b2792e06cc0cf6cf49bbb93de9a29679
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177111
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>