MB-35297: Don't call send_response twice for the same cookie
Extended testing noted that an exception was being thrown by Cookie
code and the cause was tracked down to the RangeScan being able to
call send_response twice when a scan was cancelled (depending on
the timing of the cancel).

This bug is detected by some extensions to range_scan_test and by
cleaning up the state transitions and ensuring that the cookie
is cleared once a continueScan occurs.

Change-Id: Ib12b9443d9fc00eb633767979d7608912e55d1c9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/176931
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>