Speed up STItemPagerTest/ItemPagerEvictionOrderIsSafe test by 10x
The test STItemPagerTest/ItemPagerEvictionOrderIsSafe takes ~15s to
run on macOS, DebugOptimized build.

The vast majority of this time is spent regenerating the
FailoverTables for the VBuckets as the test changes vbState. However,
the test doesn't actually to perform a 'hard' state change, one which
transfers the existing failover table (and doesn't create new entries)
is sufficient.

As such, modify the test to transfer the vb.

This reduces the overall runtime for the entire STItemPagerTest suite
from:

    1/1 Test #254: ep-engine_ep_unit_tests.EphemeralOrPersistent/STItemPagerTest ...   Passed   80.70 sec

to:

    1/1 Test #254: ep-engine_ep_unit_tests.EphemeralOrPersistent/STItemPagerTest ...   Passed   30.01 sec

Change-Id: I2069b591c836247b693871efe03265e4e620556d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177140
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>