MB-52902: Populate kvstore rev if no vbstate found
Currently we avoid loading vBucket states into the KVStore cache if
we find that the vBucket state does not exist on disk. This prevents
warmup from attempting to load a vBucket in a default (dead) state
which throws and aborts warmup for the shard. In doing so, we
currently skip the population of the kvstore revision in MagmaKVStore.
This is problematic as magma asserts that kvstore revisions only
increase and it is aware of the older revisions on disk. When we then
flush to the new revision of the vBucket magma fires an assertion as
the revision is not monotonic.

Always cache the latest kvstore revision from magma regardless of
whether there is a vBucket state in existence for the revision.

Change-Id: I920e8714f8fc4c6ee9633c4c18ce3fabcace96bf
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177346
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>