MB-52956: Update lastReadSeqno at the end of an OSO backfill
The idea behind an OSO backfill is to open a snapshot that stores many
collections, yet read only the part of the index the encompasses the
desired collection - all done using the key index.

The MB identified that if the cursor was dropped during the OSO
backfill, the next backfill is created with a start-seqno to cover
a much wider range than is required.

E.g.
An initial backfill runs and the snapshot covers 1000 keys (seqno 0 to
1000), however the target collection only occupies one item at seqno 5.
During the OSO backfill the stream's cursor is dropped.

Before the ActiveStream can switch from backfill (OSO) to in-memory
a new backfill is required to fill in the gap. A seqno backfill now
runs and opens the disk snapshot, this time the snapshot now represents
seqno 0 to 1500, here the ActiveStream sets the start-seqno of the
backfill to be seqno 6, but we know that from 0 to 1000 has been seen
by the stream and stores no data for the collection, yet the backfill
now scans from 6 to 1500.

By updating lastReadSeqno to maxSeqno in the above example the second
backfill begins at 1001.

Change-Id: If65bfa3c59310ca707ac29f53962958f35745c9e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177551
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>