Split up TTL_Expiry test
Currently the test stores an item, then attempts to expire it via
compaction. It does not wait for the item to be persisted before
the compaction is attempted so a slower running machine may run the
compaction on an empty vBucket and not expire anything. The test
also does a get on the item after the compaction. The get should
expire the item anyway.

Currently it's unclear which method expires the get. It should be
compacted, provided the item has already been persisted, but that
is not guarateed with the current test. Split it into two.

Change-Id: I6f9affebab61d16daa6b3f1b6720f1bed5ca8336
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177738
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>