MB-35297: Check for cancellation when RangeScan goes idle
The function changing to Idle considered Cancelled a bad input
state, this is incorrect as a scan can change to Cancelled whilst
it's Continuing and in the window where the scan is yielding or
at the limit (going Idle).

This commit changes the idle state path to accept Cancelled and
ensure any visible status is updated if we have been Cancelled.

Change-Id: I2df8896c3b7ad85fcf1d32ecc78404c2f0000c66
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177790
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>