Avoid calling mc_time_clock_tick from AdjustTimeOfDayExecutor
It is currently called from multiple threads when AdjustTimeOfDay
packets are executed which results in TSan reporting a data race. It
is only safe to call it from the event bast thread. To accomplish this
(nicely) we can instead pass the folly EventBase into mc_time which
makes this simple.

Change-Id: If98bd6a48526a67577378ae1b46437b0ec42b418
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177800
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>