MB-53197: Add support for ClusterConfigOnly bucket
The "ClusterConfigOnly" bucket is a special bucket
ns_server may create automatically by setting a cluster
manifest for the bucket. Clients may then connect
to the bucket and fetch the cluster manifest for the
bucket.

At some point ns_server may create a real bucket
with the same name as the bucket without having to
delete the bucket first.

Change-Id: I8b08065f5a142c4f09816550d64330e1d8fc6b69
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/177982
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>