MB-53118: Ensure Warmup disk scan obeys yield limit
As part of MB-52383, warmup tasks which read items from disk (key
and/or value scan) yield their running thread after 10ms, to allow
other tasks on the Reader thread pool to execute.

However when calculating when to yield, we reset the deadline (10ms
from now) at the start of visiting each vBucket, instead of just once
per WarmupBackfillTask::run() invocation. For datasets where a single
vBucket can be visited in less than the yield limit (10ms), we can end
up visiting multiple vBuckets before the task yields.

This manifests in "Slow runtime" log messages for WarmupBackfillTasks,
as the expected duration is the same as the yield limit:

    2022-07-28T11:49:07.779314+01:00 WARNING (default) Slow runtime for 'Warmup - key dump shard 3' on thread reader_worker_2: 22 ms

This isn't _too_ much of a problem with a yield limit of 10ms, but if
we increase the yield limit (to be more efficient when disks are
slower, and/or have few items per vBucket (so each vBucket in a shard
can be visited in 10ms) then the problem is exacerbated.

Fix by only setting the deadline once per run() invokation, so the
same "quota" of time is used for all vBuckets visited.

Change-Id: I4e3a9a0516a6f42d6352e44c1f400ae41c975e4a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178110
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>