MB-53118: Increase warmup_backfill_scan_chunk_duration to 100ms
Functional testing has highlighted that low-mid tier disks can take
over 10ms to initialise a backfill scan for warmup.

This is problematic as the current backfill chunk duration is 10s -
i.e. we will only read a single item from the scan before we yield and
pause it - and the overall warmup throughput is low.

Increase the chunk duration to 100ms - hopefully a reasonable balance
between warmup throughput, and allowing other reader threads to be
scheduled.

Change-Id: I24068732f6e081c2c6006ad0a5b3b03d72196c0e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178111
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>