Throw explicit exception if ckptMaxSize < 1
In CheckpointConfig::setCheckpointMaxSize we assert that the value
is >= 1. The calculation takes multiple parameters and the resulting
value could be a result of unexpected values in many of the
parameters. To more clearly determine why the checkpoint max size
may be computed incorrectly, throw an explicit exception in the
caller which logs all of the parameters used to calculate the value.

Change-Id: I6e2b5af9dc1ffa87b76f7531c6f7a71a3e3bef56
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178057
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>