MB-53205: Hold VBucket stateLock while calling fetchValidValue
We should always hold the vBucket state lock while calling
fetchValidValue as it can expire items. Were we to not hold the
vBucket state lock then we could begin expiring an item in the
active state and finish expiring it in the replica (/pending/dead)
state. This would result in a mismatch between active and replica
data sets.

Change-Id: I8eb893d86b6a1e3fc5505d129605d0b1a97d1a2b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178274
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>